+++ saleor/order/tests/test_order_actions_refund_products.py
@@ -62,7 +62,9 @@ def test_create_refund_fulfillment_only_order_lines(
             == current_allocation.quantity_allocated
         )
     amount = sum([line.unit_price_gross_amount * 2 for line in order_lines_to_refund])
-    mocked_refund.assert_called_once_with(payment_dummy, ANY, amount)
+    mocked_refund.assert_called_once_with(
+        payment_dummy, ANY, amount=amount, channel_slug=order_with_lines.channel.slug
+    )
     mocked_order_updated.assert_called_once_with(order_with_lines)
 
 
@@ -106,7 +108,9 @@ def test_create_refund_fulfillment_included_shipping_costs(
         assert line.quantity_unfulfilled == original_quantity.get(line.pk) - 2
     amount = sum([line.unit_price_gross_amount * 2 for line in order_lines_to_refund])
     amount += order_with_lines.shipping_price_gross_amount
-    mocked_refund.assert_called_once_with(payment_dummy, ANY, amount)
+    mocked_refund.assert_called_once_with(
+        payment_dummy, ANY, amount=amount, channel_slug=order_with_lines.channel.slug
+    )
     mocked_order_updated.assert_called_once_with(order_with_lines)
 
 
@@ -150,7 +154,9 @@ def test_create_refund_fulfillment_only_fulfillment_lines(
     amount = sum(
         [line.order_line.unit_price_gross_amount * 2 for line in fulfillment_lines]
     )
-    mocked_refund.assert_called_once_with(payment_dummy, ANY, amount)
+    mocked_refund.assert_called_once_with(
+        payment_dummy, ANY, amount=amount, channel_slug=fulfilled_order.channel.slug
+    )
     mocked_order_updated.assert_called_once_with(fulfilled_order)
 
 
@@ -193,5 +199,7 @@ def test_create_refund_fulfillment_custom_amount(
 
     for line in fulfillment_lines:
         assert line.quantity == original_quantity.get(line.pk) - 2
-    mocked_refund.assert_called_once_with(payment_dummy, ANY, amount)
+    mocked_refund.assert_called_once_with(
+        payment_dummy, ANY, amount=amount, channel_slug=fulfilled_order.channel.slug
+    )
     mocked_order_updated.assert_called_once_with(fulfilled_order)